module structure modification
